requirements.txt
Flask
python-dotenv
peewee
requests
waitress